requirements.txt
crewai[tools]==0.28.8
langchain-groq==0.1.3
exa_py
fastapi
uvicorn[standard]
streamlit
python-dotenv
requests
pydantic
tenacity
